+++ Libplanet.Tests/Blockchain/BlockChainTest.cs
@@ -705,6 +705,42 @@ namespace Libplanet.Tests.Blockchain
         }
 
         [Fact]
+        public async Task ActionRenderersHaveDistinctContexts()
+        {
+            var policy = new NullPolicy<DumbAction>();
+            var store = new DefaultStore(null);
+            var generatedRandomValueLogs = new List<int>();
+            IActionRenderer<DumbAction>[] renderers = Enumerable.Range(0, 2).Select(i =>
+                new LoggedActionRenderer<DumbAction>(
+                    new AnonymousActionRenderer<DumbAction>
+                    {
+                        ActionRenderer = (act, context, nextStates) =>
+                            // Consuming the random state through IRandom.Next() should not
+                            // affect contexts passed to other action renderers.
+                            generatedRandomValueLogs.Add(context.Random.Next()),
+                    },
+                    Log.Logger.ForContext("RendererIndex", i)
+                )
+            ).ToArray();
+            BlockChain<DumbAction> blockChain = TestUtils.MakeBlockChain(
+                policy,
+                store,
+                renderers: renderers
+            );
+            var privateKey = new PrivateKey();
+            var action = new DumbAction(default, string.Empty);
+            var actions = new[] { action };
+            blockChain.MakeTransaction(privateKey, actions);
+            Block<DumbAction> block = await blockChain.MineBlock(_fx.Address1, append: false);
+
+            generatedRandomValueLogs.Clear();
+            Assert.Empty(generatedRandomValueLogs);
+            blockChain.Append(block);
+            Assert.Equal(2, generatedRandomValueLogs.Count);
+            Assert.Equal(generatedRandomValueLogs[0], generatedRandomValueLogs[1]);
+        }
+
+        [Fact]
         public async Task RenderActionsAfterBlockIsRendered()
         {
             var policy = new NullPolicy<DumbAction>();