+++ Libplanet/Action/IActionContext.cs
@@ -1,3 +1,5 @@
+using System.Diagnostics.Contracts;
+
 namespace Libplanet.Action
 {
     /// <summary>
@@ -10,17 +12,20 @@ namespace Libplanet.Action
         /// <see cref="Address"/> of an account who made and signed
         /// a transaction that an executed <see cref="IAction"/> belongs to.
         /// </summary>
+        [Pure]
         Address Signer { get; }
 
         /// <summary>
         /// <see cref="Address"/> of a block miner account.
         /// </summary>
+        [Pure]
         Address Miner { get; }
 
         /// <summary>
         /// Block index of a transaction that an executed <see cref="IAction"/>
         /// belongs to.
         /// </summary>
+        [Pure]
         long BlockIndex { get; }
 
         /// <summary>
@@ -29,6 +34,7 @@ namespace Libplanet.Action
         /// in <see cref="PreviousStates"/>.
         /// </summary>
         /// <seealso cref="Libplanet.Tx.Transaction{T}.Create"/>
+        [Pure]
         bool Rehearsal { get; }
 
         /// <summary>
@@ -41,6 +47,7 @@ namespace Libplanet.Action
         /// instances can be returned by <see
         /// cref="IAction.Execute(IActionContext)"/> method.</para>
         /// </summary>
+        [Pure]
         IAccountStateDelta PreviousStates { get; }
 
         /// <summary>
@@ -53,5 +60,16 @@ namespace Libplanet.Action
         /// <returns>A random object that shares interface mostly equivalent
         /// to <see cref="System.Random"/>.</returns>
         IRandom Random { get; }
+
+        /// <summary>
+        /// Returns a clone of this context, except that its <see cref="Random"/> has the unconsumed
+        /// state (with the same seed).  The clone and its original are a distinct instance
+        /// each other, in other words, one's state transfer must not affect the other one
+        /// (i.e., consuming <see cref="Random"/> source should be local to a context instance).
+        /// </summary>
+        /// <returns>A clone instance, which is distinct from its original.  Its internal state
+        /// is entirely equivalent to the original's unconsumed initial state.</returns>
+        [Pure]
+        IActionContext GetUnconsumedContext();
     }
 }